.NET 6 finally compiles, but rudimentary. Need to go through, clean, check and fix a bunch of stuff.

+++ FluxSolution/BaseLibrary/Source/Types/DateTime/DayOfWeek.em.cs
@@ -11,7 +11,7 @@ namespace Flux
     {
       LocateNearestDayOfWeek(source, dayOfWeek, proper, out nearestPrevious, out nearestNext);
 
-      return new System.DateTime(System.Convert.ToInt64(source.Ticks.RoundToBoundary(mode, nearestPrevious.Ticks, nearestNext.Ticks)));
+      return new System.DateTime(System.Convert.ToInt64(((double)(source.Ticks)).RoundToBoundary(mode, (double)nearestPrevious.Ticks, (double)nearestNext.Ticks)));
     }
